refactor: split channel setup into shared flow modules
